+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/java/Java.java
@@ -67,6 +67,10 @@ public class Java extends FormatterFactory {
 		addStepFactory(importOrder);
 	}
 
+	public void addOpenJavaFormat(OpenJavaFormat openJavaFormat) {
+		addStepFactory(openJavaFormat);
+	}
+
 	public void addPalantirJavaFormat(PalantirJavaFormat palantirJavaFormat) {
 		addStepFactory(palantirJavaFormat);
 	}